+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_device_upgrade.py
@@ -11,47 +11,56 @@ from azure.cli.testsdk.scenario_tests import AllowLargeResponse
 Device tests scenarios
 """
 
-from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer
+from azure.cli.testsdk import ScenarioTest
+
 from .config import CONFIG
 
 
 def setup_scenario1(test):
-    ''' Env setup_scenario1 '''
+    """Env setup_scenario1"""
     pass
 
 
 def cleanup_scenario1(test):
-    '''Env cleanup_scenario1 '''
+    """Env cleanup_scenario1"""
     pass
 
 
 def call_scenario1(test):
-    ''' # Testcase: scenario1'''
+    """# Testcase: scenario1"""
     setup_scenario1(test)
     step_upgrade(test, checks=[])
     cleanup_scenario1(test)
 
 
 def step_upgrade(test, checks=None):
-    '''Device upgrade operation'''
+    """Device upgrade operation"""
     if checks is None:
         checks = []
     test.cmd(
-        'az networkfabric device upgrade --resource-group {upgradeDeviceRGName} --resource-name {upgradeDeviceName} --version {upgradeVersion}', checks=checks)
+        "az networkfabric device upgrade --resource-group {upgradeDeviceRGName} --resource-name {upgradeDeviceName} --version {upgradeVersion}",
+        checks=checks,
+    )
 
 
 class GA_DeviceUpgradeScenarioTest1(ScenarioTest):
-    ''' DeviceScenario test'''
+    """DeviceScenario test"""
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.kwargs.update({
-            'upgradeDeviceRGName': CONFIG.get('NETWORK_DEVICE', 'upgrade_resource_group'),
-            'upgradeDeviceName': CONFIG.get('NETWORK_DEVICE', 'upgrade_device_name'),
-            'upgradeVersion': CONFIG.get('NETWORK_DEVICE', 'upgrade_version')
-        })
+        self.kwargs.update(
+            {
+                "upgradeDeviceRGName": CONFIG.get(
+                    "NETWORK_DEVICE", "upgrade_resource_group"
+                ),
+                "upgradeDeviceName": CONFIG.get(
+                    "NETWORK_DEVICE", "upgrade_device_name"
+                ),
+                "upgradeVersion": CONFIG.get("NETWORK_DEVICE", "upgrade_version"),
+            }
+        )
 
     @AllowLargeResponse()
     def test_GA_Device_upgrade_scenario1(self):
-        ''' test scenario for Device upgrade operation'''
+        """test scenario for Device upgrade operation"""
         call_scenario1(self)